Use __objc_nlclslist to force Swift classes to be registered with ObjC.
Per Greg's feedback, this variable obviates the need for a static
constructor when the output is compiled. In interpreted mode, we still
initialize the classes by calling +load ourselves.

This is a fix-up for r3949, but we probably still need a better long-term
solution. See <rdar://problem/13154445>.

Swift SVN r3956