ABI: Fix some inaccuracies.
The production for 'nominal-type' is not an 'entity' because the type of the type is not redundantly mangled as part of the name. Our mangling can't represent nested modules yet, so update the substitutions example to talk about nested classes instead. Move the docs for known-module and known-nominal-type together with that for substitutions because they are essentially special-case substitutions.

Swift SVN r4094