Utils: Add basic pygments lexer
Treat this as something to be replaced as soon as possible; it works for the white paper, with the odd issue, but may not be so great for more complex code. Your mileage may vary.
Originally written never to be seen by anyone but me, I commit this against my better judgement; may y'all have mercy on my soul.

Swift SVN r4113