Array literals default to Slice<T>.
Introduce the array literal constraint kind and give it the default
value of 'Slice'. When this default type is used (e.g., as a type
binding in a new child system), open it so that the actual type of the
slice will be separately determined.

This means that

  var fruits = ["banana", "apple", "cherry"]

now works and gets the type String[].


Swift SVN r4136