Start tracking (uniqued) failures that occur while solving a constraint system.
For each actual error encountered while solving a constraint system,
create an object that describes that particular failure. These objects
are uniqued based on the locator, failure kind, and types/names
involved.

For now, only provide a diagnostic if there is a single failure (after
uniquing) that is a conversion constraint.



Swift SVN r4356