Decouple the solution to a constraint system from the constraint system itself.
The model of explicitly creating a tree of constraint systems and
saving the leaves that succeed is completely broken. Start detangling
it by capturing the essence of a solution in a separate class that
contains only type variable bindings and overload choices. Use that
new Solution class when comparing and applying solutions.


Swift SVN r4378