SILGen: Start implementing 'emit into' behavior.
If the result of an expression is address-only, we can store the result in-place if we have an uninitialized destination buffer for it, such as in a variable initialization or return. Modify SILGenFunction::visitExpr to write to the buffer of an Initialization if one is available instead of allocating a temporary.

A couple cases that still need to be implemented are destructuring tuple initializations (e.g. 'var (a, b) = (foo(), bar())') and apply expression results.

Swift SVN r4492