Makefiles: Fix a race in creating lib/swift.
The rules in runtime/Makefile create a lib/swift symlink without a dependency on
lib/swift, but the library rules all create a real directory, creating a race
between making the symlink and making the directory and breaking the build if
the latter wins. Make a formal rule for lib/swift that creates a symlink, and
update all the .dir dependencies on the directory to instead depend on the
symlink.



Swift SVN r4537