Solution application: build archetype member references directly.
This change uncovered a bug in the constraint-based type checker where
we didn't fully substitute the base type of an archetype member
reference through the types of the members of protocols to which that
archetype conforms. The end result is that we wouldn't properly
substitute into associated types.



Swift SVN r4651