Solution application: handle construction of struct types.
When we type-check a construction T(x), we keep track of the actual
constructor used in the type check (because it's part of the
solution). Dig that constructor out of the solution and call it
directly, rather than relying on the old type checker to do the work.


Swift SVN r4690