CMake: Fix the zillions of warnings from my last tweak.
Here's the whole story:

1. In Xcode, CMake appends the configuration name to the build path.
2. We want certain libraries to end up in lib/swift/
3. Our "lib/" is actually a symlink to <build>/lib/<configuration>/,
   so there's no need for the bit inside swift/ to include *another*
   configuration.
4. CMake's Xcode generator uses Makefiles to perform some support tasks
   in the shell, since it doesn't know how to generate shell script
   build phases.

The easiest way to accomplish (2) is to prevent (1), which is okay because
of (3). The only way to prevent (1) is to manually set per-configuration
output directories, but that causes (4) to see four identical targets,
because the "per-configuration" bit is resolved at build-time, not at
generation-time. This patch fixes that by hand-substituting the
per-configuration bit on a per-configuration basis at generation time.

Another way to accomplish (2) would be to hand-assemble the desired directory
layout after the fact, or to change our symlink hack from linking lib/ to
linking lib/swift/. One reason we didn't do that before is because we wanted
lib/clang/ to exist as well, but currently that's in lib/swift/clang because
of issues with -isysroot.

Swift SVN r4831